WW-5482 Moves locale related code into org.apache.struts2.locale
