+++ core/src/main/java/org/apache/struts2/components/I18n.java
@@ -21,14 +21,14 @@ package org.apache.struts2.components;
 import java.io.Writer;
 import java.util.ResourceBundle;
 
-import org.apache.struts2.LocaleProviderFactory;
+import org.apache.struts2.locale.LocaleProviderFactory;
 import org.apache.struts2.text.LocalizedTextProvider;
 import org.apache.struts2.text.TextProviderFactory;
 import org.apache.struts2.views.annotations.StrutsTag;
 import org.apache.struts2.views.annotations.StrutsTagAttribute;
 import org.apache.struts2.StrutsException;
 
-import org.apache.struts2.LocaleProvider;
+import org.apache.struts2.locale.LocaleProvider;
 import org.apache.struts2.text.TextProvider;
 import org.apache.struts2.inject.Inject;
 import org.apache.struts2.util.ValueStack;
