+++ core/src/main/java/org/apache/struts2/config/StrutsBeanSelectionProvider.java
@@ -22,7 +22,7 @@ import ognl.MethodAccessor;
 import org.apache.struts2.ActionProxyFactory;
 import org.apache.struts2.FileManager;
 import org.apache.struts2.FileManagerFactory;
-import org.apache.struts2.LocaleProviderFactory;
+import org.apache.struts2.locale.LocaleProviderFactory;
 import org.apache.struts2.text.LocalizedTextProvider;
 import org.apache.struts2.ObjectFactory;
 import org.apache.struts2.StrutsConstants;
