+++ core/src/main/java/org/apache/struts2/locale/DefaultLocaleProviderFactory.java
@@ -16,20 +16,13 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.struts2;
+package org.apache.struts2.locale;
 
-/**
- * Allows delegate creation of {@link LocaleProvider} to another implementation provided
- * by a user. It also allows avoid problems with too many dependencies as {@link LocaleProvider}
- * is implemented by the {@link ActionSupport} which can be defined as a bean in Spring.
- */
-public interface LocaleProviderFactory {
+public class DefaultLocaleProviderFactory implements LocaleProviderFactory {
 
-    /**
-     * Create a new instance of {@link LocaleProvider}.
-     *
-     * @return the localeProvider.
-     */
-    LocaleProvider createLocaleProvider();
+    @Override
+    public LocaleProvider createLocaleProvider() {
+        return new DefaultLocaleProvider();
+    }
 
 }
