+++ core/src/main/java/org/apache/struts2/locale/LocaleProvider.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.struts2;
+package org.apache.struts2.locale;
 
 import org.apache.commons.lang3.LocaleUtils;
 import org.apache.commons.lang3.StringUtils;
@@ -25,22 +25,18 @@ import java.util.Locale;
 
 /**
  * Indicates that the implementing class can provide its own {@link Locale}.
- *
  * <p>
  * This is useful for when an action may wish override the default locale. All that is
  * needed is to implement this interface and return your own custom locale.
  * The {@link org.apache.struts2.text.TextProvider} interface uses this interface
  * heavily for retrieving internationalized messages from resource bundles.
- * </p>
- *
- * @author Jason Carreira
  */
 public interface LocaleProvider {
 
     /**
      * Gets the provided locale.
      *
-     * @return  the locale.
+     * @return the locale.
      */
     Locale getLocale();
 
@@ -62,6 +58,7 @@ public interface LocaleProvider {
 
     /**
      * Tries to convert provided locale string into {@link Locale} or returns null
+     *
      * @param localeStr a String representing locale, e.g.: en_EN
      * @return instance of {@link Locale} or null
      * @since Struts 6.5.0
