+++ core/src/main/java/org/apache/struts2/text/TextProvider.java
@@ -18,6 +18,7 @@
  */
 package org.apache.struts2.text;
 
+import org.apache.struts2.locale.LocaleProvider;
 import org.apache.struts2.util.ValueStack;
 
 import java.util.List;
@@ -33,9 +34,9 @@ import java.util.ResourceBundle;
  * global texts to be defined for an application base class.
  *
  * <p>
- * You can override {@link org.apache.struts2.LocaleProvider#getLocale()} to change the behaviour of how
+ * You can override {@link LocaleProvider#getLocale()} to change the behaviour of how
  * to choose locale for the bundles that are returned. Typically you would
- * use the {@link org.apache.struts2.LocaleProvider} interface to get the users configured locale.
+ * use the {@link LocaleProvider} interface to get the users configured locale.
  * </p>
  *
  * <p>
@@ -56,7 +57,7 @@ import java.util.ResourceBundle;
  * </p>
  * @author Jason Carreira
  * @author Rainer Hermanns
- * @see org.apache.struts2.LocaleProvider
+ * @see LocaleProvider
  * @see TextProviderSupport
  */
 public interface TextProvider {
