+++ core/src/test/java/org/apache/struts2/text/TextProviderSupportTest.java
@@ -18,7 +18,7 @@
  */
 package org.apache.struts2.text;
 
-import org.apache.struts2.LocaleProvider;
+import org.apache.struts2.locale.LocaleProvider;
 import org.apache.struts2.XWorkTestCase;
 
 import java.util.ArrayList;
